# gensql/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package gensql

import (
	"database/sql/driver"
	"fmt"

	"github.com/jackc/pgx/v5/pgtype"
)

type Prstat string

const (
	PrstatOpen   Prstat = "open"
	PrstatMerged Prstat = "merged"
)

func (e *Prstat) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = Prstat(s)
	case string:
		*e = Prstat(s)
	default:
		return fmt.Errorf("unsupported scan type for Prstat: %T", src)
	}
	return nil
}

type NullPrstat struct {
	Prstat Prstat
	Valid  bool // Valid is true if Prstat is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullPrstat) Scan(value interface{}) error {
	if value == nil {
		ns.Prstat, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.Prstat.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullPrstat) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.Prstat), nil
}

type PullRequest struct {
	PullReqID     string
	PullReqName   string
	AuthorID      string
	PullReqStatus Prstat
	CreatedAt     pgtype.Timestamp
	MergedAt      pgtype.Timestamp
}

type ReviewersToPullRequest struct {
	UserID    string
	PullReqID string
}

type Team struct {
	TeamName string
}

type User struct {
	UserID   string
	UserName string
	IsActive bool
}

type UsersToTeam struct {
	UserID   string
	TeamName string
}